refactor: Improve functions for system; add color to Logger output.

+++ src/ame_system.cpp
@@ -30,12 +30,13 @@
 #include <format>
 #include <fstream>
 #include <optional>
+#include <regex>
 #include <string>
 
 std::optional<pid_t> FindPidByPackageName(std::string_view packageName) {
     DIR *dirp = opendir("/proc");
     if (dirp == nullptr) {
-        logger.Error("Pid not find: {}:", strerror(errno));
+        logger.Error("pid not find: {}:", strerror(errno));
         return std::nullopt;
     }
     dirent *dp = nullptr;
@@ -47,18 +48,19 @@ std::optional<pid_t> FindPidByPackageName(std::string_view packageName) {
         }
         std::string pidStr = dp->d_name;
         if (pidStr.find_first_not_of("0123456789") != std::string::npos) {
+            // not numeric name
             continue;
         }
         std::string cmdlinePath = std::format("/proc/{}/cmdline", pidStr);
-        std::fstream cmdlineFile(cmdlinePath, std::ios::in);
+        std::ifstream cmdlineFile(cmdlinePath);
         if (!cmdlineFile.is_open()) {
-            logger.Debug("Failed to open: {}:", cmdlinePath);
+            logger.Error("Failed to open: {}:", cmdlinePath);
             continue;
         }
         std::string cmdline;
-        std::getline(cmdlineFile, cmdline);
+        std::getline(cmdlineFile, cmdline, '\0');
         cmdlineFile.close();
-        if (cmdline.find(packageName) != std::string::npos) {
+        if (cmdline == packageName) {
             result = std::stoi(pidStr);
             break;
         }
@@ -67,36 +69,70 @@ std::optional<pid_t> FindPidByPackageName(std::string_view packageName) {
     return result;
 }
 
-bool IsProcessStopped(pid_t pid) {
-    return true;
+std::optional<bool> IsProcessStopped(pid_t pid) {
+    std::string statusPath = std::format("/proc/{}/status", pid);
+    std::ifstream statusFile(statusPath);
+    if (!statusFile.is_open()) {
+        logger.Error("Failed to open: {}:", statusPath);
+        return std::nullopt;
+    }
+    std::string line;
+    std::smatch matches;
+    std::regex stateRegex("^State:\\s+\\b(\\S+?)\\b\\s+\\((\\S+?)\\)");
+    while (getline(statusFile, line)) {
+        if (!std::regex_search(line, matches, stateRegex)) {
+            continue;
+        }
+        std::string stateCode = matches[1].str();
+        if (stateCode == "T") {
+            logger.Debug("Process {} is in stopped state.", pid);
+            return true;
+        } else {
+            std::string state = matches[2].str();
+            logger.Debug("Process {} is not in stopped state: {} ({}).", pid, stateCode, state);
+            return false;
+        }
+    }
+    // statusFile.close();
+    return std::nullopt;
 }
 
 bool FreezeProcess(pid_t pid) {
     if (kill(pid, SIGSTOP) == -1) {
-        logger.Error("Failed to send SIGSTOP to process {}", pid);
+        logger.Error("Failed to send SIGSTOP to process {}.", pid);
         return false;
     }
-    logger.Info("Success to send SIGSTOP to process {}", pid);
-    /* usleep(1000); // sleep 1ms, wait for process status
-    if (!IsProcessStopped(pid)) {
+    logger.Debug("Success to send SIGSTOP to process {}.", pid);
+    usleep(1000 * 10); // sleep 10ms, wait for process status
+    auto isStopped = IsProcessStopped(pid);
+    if (!isStopped.has_value()) {
+        logger.Error("Failed to get state of process {}.", pid);
         return false;
-    } */
-    logger.Info("Process {} freeze success.", pid);
-    return true;
+    }
+    if (*isStopped) {
+        logger.Debug("Process {} froze successfully.", pid);
+        return true;
+    }
+    logger.Debug("Failed to freeze process {}.", pid);
+    return false;
 }
 
 bool TryToResumeProsess(pid_t pid, int attempts) {
-    for (int i = 0; i < attempts; ++i) {
+    if (attempts < 1) {
+        logger.Error("attempts less than one: {}.", attempts);
+        return false;
+    }
+    for (int i = 1, j = 0; i <= attempts; ++i) {
         if (kill(pid, SIGCONT) == -1) {
+            logger.Debug("Failed to send SIGCONT to process {} for {} time.", pid, ++j);
             continue;
         }
         usleep(1000); // sleep 1ms, wait for process status
-        /* if (IsProcessStopped(pid)) {
-            usleep(1000 * 100); // sleep 100ms, wait for another try
-            continue;
-        } */
-        logger.Info("Process {} resumed successfully.", pid);
-        return true;
+        if (auto isStop = IsProcessStopped(pid); isStop.has_value() && !*isStop) {
+            logger.Debug("Process {} resumed successfully.", pid);
+            return true;
+        }
+        usleep(1000 * 100); // sleep 100ms, wait for another try
     }
     logger.Error("Failed to resume process {} after {} attempts.", pid, attempts);
     return false;
